refactored plugin to conform to package diagram
