+++ plugins/essif-lab/src/Application/Controllers/Activate.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace TNO\Essiflab\Application\Controllers;
+
+defined('ABSPATH') or die();
+
+use TNO\EssifLab\Application\Workflows\Constructors\CoreAbstract;
+
+class Activate extends CoreAbstract
+{
+}
