+++ plugins/essif-lab/src/Application/Workflows/Constructors/CoreInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace TNO\EssifLab\Contracts;
+namespace TNO\EssifLab\Application\Workflows\Constructors;
 
 defined('ABSPATH') or die();
