--- plugins/essif-lab/src/Controllers/Activate.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace TNO\EssifLab\Controllers;
-
-defined('ABSPATH') or die();
-
-use TNO\EssifLab\Extendables\CoreAbstract;
-
-class Activate extends CoreAbstract
-{
-}
